requirements.txt
langchain==0.0.275
neo4j==5.28.1
paramiko==3.5.1
pydantic==2.11.4
python-dotenv==1.1.0
rdkit==2024.9.5
Requests==2.32.3
rmrkl==0.0.3
openai==0.27.8
PyYAML==6.0.2
nest-asyncio==1.6.0
biopython==1.84
pandas==2.2.2
# MCP Python client (Model Context Protocol)
mcp>=1.12.0
# biomcp CLI for BioMCP server
biomcp-python
